%p OK

+++ ft_itoh.c
@@ -6,17 +6,17 @@
 /*   By: bledda <bledda@student.42nice.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/07 20:01:58 by bledda            #+#    #+#             */
-/*   Updated: 2021/04/09 13:48:08 by bledda           ###   ########.fr       */
+/*   Updated: 2021/04/12 18:54:18 by bledda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-char	*ft_itoh(unsigned int nb)
+char	*ft_itoh(unsigned long long nb)
 {
 	char			itoh[100];
 	int				i;
-	unsigned int	save_nb;
+	unsigned long long save_nb;
 	char			*r;
 
 	i = 0;
